Fix out of bounds reads and small issues found in a review
- e_them.c: the enemy randomizer read the high word of e_them_rndseed
  2 bytes past the variable; read the variable's own high word.
- scr_imain.c: the hall of fame printed the names with %s, but they are
  not null terminated, and the U32 score with %d. Copy the names with
  their fixed size instead.
- hiscores.c: only the first 8 valid lines of highscores.txt were used,
  even if better scores followed. Keep the 8 best entries instead.
- Fix compiler warnings: missing parentheses, a signed/unsigned
  comparison, a possibly uninitialized variable and size_t printed
  with %u in the memory debug output.

+++ source/xrick/system/sysarg_sdl.c
@@ -173,35 +173,35 @@ static bool sysarg_scankeys(const char *keys)
   i = 0;
 
   j = 0;
-  while (keys[i] != '\0' && keys[i] != '-' && j + 1 < sizeof k) k[j++] = keys[i++];
+  while (keys[i] != '\0' && keys[i] != '-' && j + 1 < (int)sizeof k) k[j++] = keys[i++];
   if (keys[i++] == '\0') return false;
   k[j] = '\0';
   syskbd_left = sysarg_sdlcode(k);
   if (!syskbd_left) return false;
 
   j = 0;
-  while (keys[i] != '\0' && keys[i] != '-' && j + 1 < sizeof k) k[j++] = keys[i++];
+  while (keys[i] != '\0' && keys[i] != '-' && j + 1 < (int)sizeof k) k[j++] = keys[i++];
   if (keys[i++] == '\0') return false;
   k[j] = '\0';
   syskbd_right = sysarg_sdlcode(k);
   if (!syskbd_right) return false;
 
   j = 0;
-  while (keys[i] != '\0' && keys[i] != '-' && j + 1 < sizeof k) k[j++] = keys[i++];
+  while (keys[i] != '\0' && keys[i] != '-' && j + 1 < (int)sizeof k) k[j++] = keys[i++];
   if (keys[i++] == '\0') return false;
   k[j] = '\0';
   syskbd_up = sysarg_sdlcode(k);
   if (!syskbd_up) return false;
 
   j = 0;
-  while (keys[i] != '\0' && keys[i] != '-' && j + 1 < sizeof k) k[j++] = keys[i++];
+  while (keys[i] != '\0' && keys[i] != '-' && j + 1 < (int)sizeof k) k[j++] = keys[i++];
   if (keys[i++] == '\0') return false;
   k[j] = '\0';
   syskbd_down = sysarg_sdlcode(k);
   if (!syskbd_down) return false;
 
   j = 0;
-  while (keys[i] != '\0' && keys[i] != '-' && j + 1 < sizeof k) k[j++] = keys[i++];
+  while (keys[i] != '\0' && keys[i] != '-' && j + 1 < (int)sizeof k) k[j++] = keys[i++];
   if (keys[i] != '\0') return false;
   k[j] = '\0';
   syskbd_fire = sysarg_sdlcode(k);
