+++ source/xrick/system/sysmem_sdl.c
@@ -60,7 +60,7 @@ void sysmem_shutdown(void)
     }
 
     IFDEBUG_MEMORY(
-        sys_printf("xrick/memory: max memory usage was %u bytes\n", maxUsedMemory);
+        sys_printf("xrick/memory: max memory usage was %lu bytes\n", (unsigned long)maxUsedMemory);
     );
 
     isMemoryInitialised = false;
@@ -90,7 +90,7 @@ void *sysmem_push(size_t size)
     stackSize += neededSize;
 
     IFDEBUG_MEMORY(
-        sys_printf("xrick/memory: allocated %u bytes\n", neededSize);
+        sys_printf("xrick/memory: allocated %lu bytes\n", (unsigned long)neededSize);
         if (stackSize > maxUsedMemory) maxUsedMemory = stackSize;
     );
 
@@ -128,7 +128,7 @@ void sysmem_pop(void * alignedPtr)
     );
 
     IFDEBUG_MEMORY(
-        sys_printf("xrick/memory: released %u bytes\n", allocatedSize);
+        sys_printf("xrick/memory: released %lu bytes\n", (unsigned long)allocatedSize);
     );
 }
